+++ src/App.tsx
@@ -2,6 +2,7 @@ import { Titlebar } from "./components/Titlebar";
 import { VellumShell } from "./components/VellumShell";
 import { VellumProvider } from "./state/vellum";
 import { ActiveEditorProvider } from "./state/activeEditor";
+import { UpdaterProvider } from "./state/updater";
 import { useWindowMaximized } from "./components/useWindowMaximized";
 
 function App() {
@@ -11,7 +12,9 @@ function App() {
       <Titlebar title="Vellum" maximized={maximized} />
       <VellumProvider>
         <ActiveEditorProvider>
-          <VellumShell />
+          <UpdaterProvider>
+            <VellumShell />
+          </UpdaterProvider>
         </ActiveEditorProvider>
       </VellumProvider>
     </div>